Recursively resolve symlinks #43
This works by using `readlink` (without the `-f` option, which is
not availble on macOS's built-in `readlink` implementation) to
resolve terminal symlinks, and Bash's `cd`'s `-P` option to resolve
all directory symlinks.

Thus this works on all modern Linux distros and macOS versions,
with any (known) combination of symlinked files and directories.